backend/requirements.txt
fastapi==0.68.1
uvicorn==0.15.0
sqlalchemy==1.4.23
psycopg2-binary==2.9.1
python-dotenv==0.19.0
openai==0.27.0
pydantic==1.8.2
python-multipart==0.0.5
aiohttp>=3.8.3,<4.0.0
numpy==1.21.2
pandas==1.3.3
scikit-learn==0.24.2
langchain==0.0.335
pgvector==0.2.3
tqdm==4.67.1
alembic==1.7.7 